+++ packages/vue/src/composables/useI18n.ts
@@ -1,14 +1,30 @@
 import tLocale from './t'
+import dnLocale from './dn'
+import vLocale from './v'
+import uLocale from './u'
 import useLocale from './useLocale'
 
 function useI18n() {
-  const locale = useLocale().value
+  const localeRef = useLocale();
+  const locale = localeRef.value
 
-  const t: typeof tLocale = (text, values, optionLocale) => {
-    return tLocale(text, values, optionLocale || locale)
+  const t: typeof tLocale = (text, values, options) => {
+    return tLocale(text, values, options || { locale })
   }
 
-  return { t, locale }
+  const dn: typeof dnLocale = (code, intlOptions, options) => {
+    return dnLocale(code, intlOptions, options || { locale })
+  }
+
+  const v: typeof vLocale = (value, options) => {
+    return vLocale(value, options || { locale })
+  }
+
+  const u: typeof uLocale = (url, values, options) => {
+    return uLocale(url, values, options || { locale })
+  }
+
+  return { locale: localeRef, t, dn, v, u }
 }
 
 export default useI18n