+++ packages/vue/src/keys.ts
@@ -1,3 +1,3 @@
 import type { InjectionKey, Ref } from 'vue'
 
-export const localeKey = Symbol('Locale') as InjectionKey<Readonly<Ref<string>>>
+export const localeKey = Symbol('Locale') as InjectionKey<Ref<string>>