+++ packages/vue/CHANGELOG.md
@@ -4,6 +4,24 @@ name: Changelog
 description: All notable changes to this project will be documented in this file.
 ---
 
+## [Unreleased]
+
+### Added
+
+- Added `dn()`
+- Added `u()`
+- Added `v()`
+
+### Changed
+
+- Added `dn()` and `u()` to `useI18n` return
+- Changed `t()` options to be `{ locale }`
+- Changed `useLocale` to return locale ref
+
+### Fixed
+
+- Provided `locale` as not read-only `Ref`
+
 ## [0.1.0] - 2024-03-17
 
 ### Added